rename embed.py to base.py
